This branch has removed unused code and restructured the motor control into a library to be consistent with the sensors and control system code
500 lines of unused base code removed.
The motors code is now in a library. (needs checking and commenting).
Need to think about open loop and wiener filter.
Finite state machine needs to be checked over for logical errors.
e.g RunMode never gets updated, only nextMode.
It might be worth splitting the state of run mode out into functions that same way this was done in the base code.