Finished sensors code
Removed all other folders from the trimmed code. Finished the sensors code. Added a scaling factor to the motors code.  Added some comments for readability to the main.ino.

There are some potential issuses that we need to discus including when parallel is in error state and smooth operation of the align in the follow state vs initial allignment.